Store Cleanup, Ship Customization Progress, Menus, and Stats
Another pretty big update! Really working hard on this store. Major
feature to note is the developer mode under the new menu. It gives you
tons of cash and removes persistance, so feel free to die with it on.
Just for testing the menus and store without mining for days. Check out
miningBelt_Time for more details on the changes happening! Hope you
enjoy <3