Final JS Update
Dear faithful miningBelt followers! (Talking to you RJ and my Bro <3). The .exe has been updated! Ship customization is in and it is pretty sweet! (Thanks Matt!). Thats it!! miningBelt is being completely re-built from the ground up! It needs it. The first iteration was fun but just isn't worth moving forwards with anymore! I will be converting the whole project to C#, and renovating many aspects of the game. I will be making a new branch for GitHub so stay tuned! Thanks for the love, miningBelt will be something someday, promise.